fix: update event listeners for textinput
- refactor components
