+++ packages/filterpane/src/components/FilterSelector/FilterCategory.style.tsx
@@ -1,0 +1,15 @@
+import { createStyles, makeStyles } from '@equinor/fusion-react-styles';
+
+export const useStyles = makeStyles(() =>
+  createStyles({
+    FilterCategoryContainer: {
+      display: 'flex',
+      alignItems: 'center',
+    },
+    CategoryLabel: {
+      cursor: 'pointer',
+    },
+  })
+);
+
+export default useStyles;
