+++ packages/filterpane/src/components/GeneralBar/GeneralBar.style.tsx
